+++ src/Sulu/Bundle/MediaBundle/DependencyInjection/SuluMediaExtension.php
@@ -11,6 +11,7 @@
 
 namespace Sulu\Bundle\MediaBundle\DependencyInjection;
 
+use Sulu\Bundle\MediaBundle\Entity\Collection;
 use Sulu\Bundle\MediaBundle\Media\Exception\FileVersionNotFoundException;
 use Sulu\Bundle\MediaBundle\Media\Exception\FormatNotFoundException;
 use Sulu\Bundle\MediaBundle\Media\Exception\FormatOptionsMissingParameterException;
@@ -85,6 +86,9 @@ class SuluMediaExtension extends Extension implements PrependExtensionInterface
     {
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
+
+        // collection-class
+        $container->setParameter('sulu.model.collection.class', Collection::class);
 
         // image-formats
         $container->setParameter('sulu_media.image_format_files', $config['image_format_files']);