+++ CHANGELOG.md
@@ -2,6 +2,9 @@ CHANGELOG for Sulu
 ==================
 
 * dev-release/1.5
+    * HOTFIX      #4058 [ContactBundle]         Added escaping of name in contact-list
+    * HOTFIX      #4058 [ContactBundle]         Added escaping for address and bank-accounts
+    * HOTFIX      #4058 [MediaBundle]           Added escaping of name in media-list
     * HOTFIX      #4057 [SecurityBundle]        Removed password-hash from user-api
 
 * 1.5.16 (2018-06-29)