+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-release/1.5
     * HOTFIX      #4056 [MediaBundle]           Added security-check for collection permission to media-controller
+    * HOTFIX      #5060 [MediaBundle]           Added escaping of media edit overlay title
     * HOTFIX      #4058 [ContactBundle]         Added escaping of name in contact-list
     * HOTFIX      #4058 [ContactBundle]         Added escaping for address and bank-accounts
     * HOTFIX      #4058 [MediaBundle]           Added escaping of name in media-list